refactor:move static methods to utils
